Moved the navigation into a separate component for better modularity and reusability
