+++ src/components/Nav/ButtonHome/style.scss
@@ -4,6 +4,3 @@
     display: flex;
     justify-content: flex-end;     
 }
-
-
-
