+++ src/components/Nav/MobileMenu/style.scss
@@ -1,0 +1,6 @@
+@use '../../../styles/settings/colors' as v;
+
+.c-button-header {
+    display: flex;
+    justify-content: flex-end;     
+}
