+++ src/layout/AuthenticatedLayout/index.jsx
@@ -1,58 +1,23 @@
 import PropTypes from "prop-types"
-import { Link, useNavigate } from 'react-router-dom' 
-import {useDispatch, useSelector } from 'react-redux'
-import  { useEffect } from 'react'
 
 import Header from '../../components/Header'
 import Footer from '../../components/Footer'
-import UserCard from "../../components/Cards/UserCard"
-import { getIsLogged, getUserId, getUserOrganizationId } from "../../redux/selectors/user"
-import {logout}  from "../../redux/reducers/user"
+import DesktopMenu from "../../components/Nav/DesktopMenu"
 
-import { List, ListItem, ListItemButton, ListItemIcon, ListItemText } from '@mui/material'
-import { Toolbar, Divider, Box, Drawer } from '@mui/material'
-import AdminPanelSettingsIcon from '@mui/icons-material/AdminPanelSettings'
-import ContactMailIcon from '@mui/icons-material/ContactMail'
-import LogoutIcon from '@mui/icons-material/Logout'
-import ManageAccountsIcon from '@mui/icons-material/ManageAccounts'
-import ForumIcon from '@mui/icons-material/Forum'
-import PersonIcon from '@mui/icons-material/Person';
+import { Toolbar, Box, Drawer, List } from '@mui/material'
+
 import './style.scss'
-import { cleanOrganizationState } from "../../redux/reducers/organization"
+import UserCard from "../../components/Cards/UserCard"
+
+
 
 
 const drawerWidth = 240;
 
 
-
-
-
 function AuthenticatedLayout({children}) {
 
-    const dispatch = useDispatch();
-    const isLog = useSelector(getIsLogged);
-    const navigate = useNavigate();
-    const organizationId = useSelector(getUserOrganizationId);
-    const userId = useSelector(getUserId);
-
-    const handleLogout = () => {
-        dispatch(logout());
-        dispatch(cleanOrganizationState());
-    }
-
-    const data = [
-        { text: "Flux d'activité", index: <ForumIcon />, route: `/${organizationId}` },
-        { text: 'Mon profil', index: <PersonIcon/>, route: `/${organizationId}/user/${userId}`},
-        { text: 'Editer mon profil', index: <ManageAccountsIcon/>, route: `/${organizationId}/user/${userId}/edit`},
-        { text: 'Administration', index: <AdminPanelSettingsIcon />, route: `/${organizationId}/admin/members`},
-        { text: 'Contact', index: <ContactMailIcon />, route: "/about" },
     
-    ]
-    useEffect(() => {
-        if (!isLog) {
-            navigate('/');  
-        }
-    }, [isLog, navigate]);
 
     return (
         
@@ -77,35 +42,7 @@ function AuthenticatedLayout({children}) {
                             <Box className='c-box-avatar'>
                                 <UserCard/>
                             </Box>
-                            <Divider/>
-                            {data.map(({text, index, route}) => (
-                                <ListItem 
-                                    key={text}
-                                    component={Link} to={route} 
-                                    disablePadding
-                                >
-                                    <ListItemButton>
-                                        <ListItemIcon>
-                                            {index}
-                                        </ListItemIcon>
-                                        <ListItemText primary={text} />
-                                    </ListItemButton>
-                                </ListItem>
-                            ))}  
-                            <ListItem
-                                key={"Déconnexion"}
-                                disablePadding
-                                component={Link}
-                                to = "/"
-                                onClick={handleLogout}
-                            >
-                                <ListItemButton>
-                                    <ListItemIcon>
-                                        <LogoutIcon />
-                                    </ListItemIcon>
-                                    <ListItemText primary={"Déconnexion"} />
-                                </ListItemButton>
-                            </ListItem>
+                            <DesktopMenu/>
                         </List> 
                     </Box>
                 </Drawer>
